@@ -11,7 +11,7 @@
     "@turf/turf": "^5.1.6",
     "axios": "^0.21.1",
     "axios-progress-bar": "^1.2.0",
-    "bootstrap": "4.3.1",
+    "bootstrap": "5.2.2",
     "buffer": "^6.0.3",
     "date-fns": "^1.30.1",
     "geodesy": "^2.0.1",
@@ -22,13 +22,13 @@
     "mb-isochrone": "^0.1.0",
     "pluralize": "^8.0.0",
     "prop-types": "^15.7.2",
-    "react": "^16.12.0",
-    "react-bootstrap": "1.0.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "2.5.0",
     "react-collapsible": "^2.6.2",
     "react-css-loaders": "^0.0.5",
     "react-datepicker": "^4.7.0",
     "react-debounce-render": "^6.0.0",
-    "react-dom": "^16.12.0",
+    "react-dom": "^18.2.0",
     "react-error-boundary": "^1.2.5",
     "react-fast-compare": "^2.0.4",
     "react-flip-toolkit": "^7.0.13",
@@ -38,7 +38,7 @@
     "react-jsonschema-form-bs4": "^1.7.1",
     "react-mapbox-gl": "4.8.3",
     "react-redux": "^7.0.3",
-    "react-router-dom": "^6.3.0",
+    "react-router-dom": "^6.4.2",
     "react-scripts": "5.0.0",
     "react-select": "^3.0.3",
     "react-toastify": "^5.4.0",
