@@ -1,95 +1,109 @@
-import React, { useState, useEffect } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import Button from 'react-bootstrap/Button';
 import ListGroup from 'react-bootstrap/ListGroup';
-import Popover from 'react-bootstrap/Popover';
 import OverlayTrigger from 'react-bootstrap/OverlayTrigger';
-import {  DEFAULT_EVENT_SORT, SORT_DIRECTION } from '../utils/event-filter';
+import Popover from 'react-bootstrap/Popover';
 
-import styles from './styles.module.scss';
 import { ReactComponent as ArrowDown } from '../common/images/icons/arrow-down.svg';
 import { ReactComponent as ArrowUp } from '../common/images/icons/arrow-up.svg';
-import { ReactComponent as SortLines } from '../common/images/icons/sort-lines.svg';
 import { ReactComponent as CheckIcon } from '../common/images/icons/check.svg';
+import { ReactComponent as SortLines } from '../common/images/icons/sort-lines.svg';
 
-const { Title, Content } = Popover;
-const { Item } = ListGroup;
-const UP = SORT_DIRECTION.up;
-const DOWN = SORT_DIRECTION.down;
+import { DEFAULT_EVENT_SORT, SORT_DIRECTION } from '../utils/event-filter';
 
-const ColumnSort = (props) => {
+import styles from './styles.module.scss';
 
-  const { className = '', sortOptions, orderOptions, value, onChange } = props;
-  const [isSortUp, setSortDirection] = useState(value === UP);
+const ColumnSort = ({ className = '', sortOptions, orderOptions, value, onChange }) => {
+  const [isSortUp, setSortDirection] = useState(value === SORT_DIRECTION.up);
+
   const [direction, selectedOption] = value;
   const isSortModified = !DEFAULT_EVENT_SORT.includes(selectedOption);
 
-  const onClickSortOption = (option) => {
+  const onClickSortOption = useCallback((option) => {
     if (option.value !== selectedOption.value){
       onChange([direction, option]);
     }
-  };
+  }, [direction, onChange, selectedOption.value]);
 
-  const toggleSortDirection = () => {
-    const newDir = direction === UP ? DOWN : UP;
+  const toggleSortDirection = useCallback(() => {
+    const newDir = direction === SORT_DIRECTION.up ? SORT_DIRECTION.down : SORT_DIRECTION.up;
+
     onChange([newDir, selectedOption]);
-  };
+  }, [direction, onChange, selectedOption]);
 
-  const changeSortDirection = (value) => {
+  const changeSortDirection = useCallback((value) => {
     if (direction !== value){
       onChange([SORT_DIRECTION[value], selectedOption]);
     }
-  };
+  }, [direction, onChange, selectedOption]);
 
   useEffect(() => {
-    setSortDirection(direction === UP);
+    setSortDirection(direction === SORT_DIRECTION.up);
   }, [direction]);
 
   const sortDirectionIcon = isSortUp ? <ArrowUp /> : <ArrowDown />;
 
   const SortPopover = <Popover className={styles.sortPopover} id='sort-popover'>
-    <Title>
+    <Popover.Header>
       <div className={styles.popoverTitle}>
         Sort by:
       </div>
-    </Title>
-    <Content className="styles.sortList">
+    </Popover.Header>
+
+    <Popover.Body className="styles.sortList">
       <ListGroup className="styles.listGroup" data-testid="sort-options">
-        {sortOptions.map(option => <Item action className={styles.listItem} key={option.value} onClick={() => onClickSortOption(option)}>
+        {sortOptions.map((option) => <ListGroup.Item
+          action
+          className={styles.listItem}
+          key={option.value}
+          onClick={() => onClickSortOption(option)}
+        >
           {option.value === selectedOption.value && <CheckIcon className={styles.checkIcon}/>}
           {option.label}
-        </Item>)}
-
+        </ListGroup.Item>)}
       </ListGroup>
+
       <ListGroup className="styles.listGroup" data-testid="order-options">
-        {orderOptions.map(option => (
-          <Item
-            action
-            className={styles.listItem}
-            key={option.value}
-            onClick={() => changeSortDirection(option.value)}
-          >
-            {option.value === direction && <CheckIcon className={styles.checkIcon}/>}
-            {option.label}
-          </Item>
-        ))}
-
+        {orderOptions.map(option => <ListGroup.Item
+          action
+          className={styles.listItem}
+          key={option.value}
+          onClick={() => changeSortDirection(option.value)}
+        >
+          {option.value === direction && <CheckIcon className={styles.checkIcon}/>}
+          {option.label}
+        </ListGroup.Item>)}
       </ListGroup>
-    </Content>
+    </Popover.Body>
   </Popover>;
 
   return <span className={`${styles.columnWrapper} ${className}`}>
-    <OverlayTrigger shouldUpdatePosition={true} rootClose trigger='click' placement='bottom' overlay={SortPopover} flip={true} rootCloseEvent='click'>
-      <Button size='sm' variant={isSortModified ? 'primary' : 'light'} className={styles.popoverTrigger} data-testid="sort-popover-trigger">
-        <SortLines/>
-        <span> {selectedOption.label} </span>
+    <OverlayTrigger
+      shouldUpdatePosition={true}
+      rootClose
+      trigger='click'
+      placement='bottom'
+      overlay={SortPopover}
+      flip={true}
+      rootCloseEvent='click'
+    >
+      <Button
+        className={styles.popoverTrigger}
+        data-testid="sort-popover-trigger"
+        size='sm'
+        variant={isSortModified ? 'primary' : 'light'}
+      >
+        <SortLines />
+        <span>{selectedOption.label}</span>
       </Button>
     </OverlayTrigger>
+
     <Button
-      size='sm'
-      data-testid='sort-direction-toggle'
       className={styles.sortDirection}
-      variant={isSortUp ? 'primary' : 'light'}
+      data-testid='sort-direction-toggle'
       onClick={toggleSortDirection}
+      size='sm'
+      variant={isSortUp ? 'primary' : 'light'}
     >
       {sortDirectionIcon}
     </Button>
