@@ -1,30 +1,38 @@
-import React, { forwardRef, memo, useCallback, useContext, useMemo, useState, useRef, Fragment } from 'react';
-import PropTypes from 'prop-types';
-import TimeAgo from '../TimeAgo';
-import Popover from 'react-bootstrap/Popover';
+import React, { forwardRef, memo, useCallback, useContext, useMemo, useState, useRef } from 'react';
 import Overlay from 'react-bootstrap/Overlay';
-
-import { FormDataContext } from './context';
-
-import { BREAKPOINTS } from '../constants';
-import { useMatchMedia } from '../hooks';
-
-import InlineEditable from '../InlineEditable';
-import HamburgerMenuIcon from '../HamburgerMenuIcon';
-import DateTime from '../DateTime';
+import Popover from 'react-bootstrap/Popover';
+import PropTypes from 'prop-types';
 
 import { analyticsMetadata } from '../proptypes';
-
+import { BREAKPOINTS } from '../constants';
+import { FormDataContext } from './context';
 import { trackEvent } from '../utils/analytics';
+import { useMatchMedia } from '../hooks';
+
+import DateTime from '../DateTime';
+import HamburgerMenuIcon from '../HamburgerMenuIcon';
+import InlineEditable from '../InlineEditable';
+import TimeAgo from '../TimeAgo';
 
 import styles from './styles.module.scss';
 
-const EditableItemHeader = (props) => {
-  const { analyticsMetadata, readonly, afterMenuToggle, children, priority, icon: Icon, menuContent: MenuContent, title: titleProp, onTitleChange, maxTitleLength } = props;
-
+const EditableItemHeader = ({
+  afterMenuToggle,
+  analyticsMetadata,
+  children,
+  icon: Icon,
+  maxTitleLength,
+  menuContent: MenuContent,
+  onTitleChange,
+  priority,
+  readonly,
+  title: titleProp,
+}) => {
   const data = useContext(FormDataContext);
+
   const menuRef = useRef(null);
   const historyRef = useRef(null);
+
   const [headerPopoverOpen, setHeaderPopoverState] = useState(false);
   const [historyPopoverOpen, setHistoryPopoverState] = useState(false);
   const [editingTitle, setTitleEditState] = useState(false);
@@ -33,23 +41,16 @@ const EditableItemHeader = (props) => {
 
   const popoverPlacement = isExtraLargeLayout ? 'auto' : 'left';
 
+  const startTitleEdit = useCallback(() => setTitleEditState(true), []);
 
-  const startTitleEdit = useCallback(() => {
-    setTitleEditState(true);
-  }, []);
-
-  const cancelTitleEdit = useCallback(() => {
-    setTitleEditState(false);
-  }, []);
+  const cancelTitleEdit = useCallback(() => setTitleEditState(false), []);
 
   const onSaveTitle = useCallback((value) => {
     onTitleChange(value);
     setTitleEditState(false);
   }, [onTitleChange]);
 
-  const title = useMemo(() => {
-    return titleProp || data.title;
-  }, [data.title, titleProp]);
+  const title = useMemo(() => titleProp || data.title, [data.title, titleProp]);
 
   const updateTime = useMemo(() => {
     if (data.updates?.length) {
@@ -61,9 +62,7 @@ const EditableItemHeader = (props) => {
   }, [data.updated_at, data.created_at, data.updates]);
 
   const handleEscapePress = (event) => {
-    const { key } = event;
-    if (key === 'Escape'
-    && (headerPopoverOpen || historyPopoverOpen)) {
+    if (event.key === 'Escape' && (headerPopoverOpen || historyPopoverOpen)) {
       event.preventDefault();
       event.stopPropagation();
       setHeaderPopoverState(false);
@@ -74,85 +73,116 @@ const EditableItemHeader = (props) => {
   const onHamburgerMenuIconClick = () => {
     setHeaderPopoverState(!headerPopoverOpen);
     afterMenuToggle && afterMenuToggle(!headerPopoverOpen);
-    !!analyticsMetadata && trackEvent(analyticsMetadata.category, `${headerPopoverOpen?'Close':'Open'} Hamburger Menu${!!analyticsMetadata.location ? ` for ${analyticsMetadata.location}` : ''}`);
+    !!analyticsMetadata && trackEvent(
+      analyticsMetadata.category,
+      `${headerPopoverOpen?'Close':'Open'} Hamburger Menu${!!analyticsMetadata.location ? ` for ${analyticsMetadata.location}` : ''}`
+    );
   };
-
 
   const onHistoryClick = () => {
     setHistoryPopoverState(!historyPopoverOpen);
-    !!analyticsMetadata && trackEvent(analyticsMetadata.category, `${historyPopoverOpen?'Close':'Open'} History${!!analyticsMetadata.location ? ` for ${analyticsMetadata.location}` : ''}`);
+    !!analyticsMetadata && trackEvent(
+      analyticsMetadata.category,
+      `${historyPopoverOpen?'Close':'Open'} History${!!analyticsMetadata.location ? ` for ${analyticsMetadata.location}` : ''}`
+    );
   };
 
-  const HistoryLink = data.updates && <Fragment>
+  const HistoryLink = data.updates && <>
     <span ref={historyRef} onClick={onHistoryClick} className={styles.history}>
       {data.updates.length > 1 ? 'Updated' : 'Created'} <TimeAgo date={updateTime}/>
     </span>
-  </Fragment>;
+  </>;
 
-  const HistoryPopover = forwardRef((props, ref) => <Popover {...props} ref={ref} className={styles.historyPopover}> {/* eslint-disable-line react/display-name */}
-    <Popover.Title>History</Popover.Title>
-    <Popover.Content>
+  // eslint-disable-next-line react/display-name
+  const HistoryPopover = forwardRef((props, ref) => <Popover {...props} ref={ref} className={styles.historyPopover}>
+    <Popover.Header>History</Popover.Header>
+
+    <Popover.Body>
       <ul>
         {data.updates && data.updates.map((update) =>
           <li className={styles.listItem} key={update.time}>
             <div className={styles.historyItem}>
               <div className={styles.historyDetails}>
                 <div className={styles.historyMessage}>{update.message.replace(/ by [ \w+\b]*$/g, '')}</div>
-                {update?.user?.first_name && <div className={styles.historyUser}>{`${update.user.first_name} ${update.user.last_name}`.trim()}</div>}
+                {update?.user?.first_name && <div className={styles.historyUser}>
+                  {`${update.user.first_name} ${update.user.last_name}`.trim()}
+                </div>}
               </div>
+
               <DateTime className={styles.historyDate} date={update.time}/>
             </div>
           </li>
         )}
       </ul>
-    </Popover.Content>
+    </Popover.Body>
   </Popover>);
 
   return <div className={`${styles.formHeader} ${styles[`priority-${priority}`]} ${readonly ? styles.readonly : ''}`} onKeyDown={handleEscapePress}>
     <h4>
-      {!!Icon &&
-        <span className={styles.headerIcon}>
-          {Icon}
-        </span>
-      }
+      {!!Icon && <span className={styles.headerIcon}>{Icon}</span>}
+
       {data.serial_number && <span>{data.serial_number}</span>}
-      <InlineEditable editing={editingTitle} onClick={startTitleEdit} onChange={onTitleChange} onEsc={cancelTitleEdit} onCancel={cancelTitleEdit} value={title} onSave={onSaveTitle} maxLength={maxTitleLength} />
+
+      <InlineEditable
+        editing={editingTitle}
+        onClick={startTitleEdit}
+        onChange={onTitleChange}
+        onEsc={cancelTitleEdit}
+        onCancel={cancelTitleEdit}
+        value={title}
+        onSave={onSaveTitle}
+        maxLength={maxTitleLength}
+      />
+
       {children}
+
       <div className={styles.headerDetails}>
-        {!!MenuContent && <Fragment>
+        {!!MenuContent && <>
           <HamburgerMenuIcon ref={menuRef} isOpen={headerPopoverOpen} onClick={onHamburgerMenuIconClick} />
-          <Overlay show={headerPopoverOpen} target={menuRef.current} shouldUpdatePosition={true} rootClose
-            onHide={() => setHeaderPopoverState(false)} placement={popoverPlacement} trigger='click'>
+          <Overlay
+            show={headerPopoverOpen}
+            target={menuRef.current}
+            shouldUpdatePosition={true}
+            rootClose
+            onHide={() => setHeaderPopoverState(false)}
+            placement={popoverPlacement}
+            trigger='click'
+          >
             {MenuContent}
           </Overlay>
-        </Fragment>
+        </>
         }
+
         {HistoryLink}
-        <Overlay show={historyPopoverOpen} target={historyRef.current} shouldUpdatePosition={true} rootClose
-          onHide={() => setHistoryPopoverState(false)} placement={popoverPlacement} trigger='click'>
+
+        <Overlay
+          show={historyPopoverOpen}
+          target={historyRef.current}
+          shouldUpdatePosition={true}
+          rootClose
+          onHide={() => setHistoryPopoverState(false)}
+          placement={popoverPlacement}
+          trigger='click'
+        >
           <HistoryPopover />
         </Overlay>
+
         {data.state === 'resolved' && <small>resolved</small>}
       </div>
     </h4>
   </div>;
 };
 
-export default memo(EditableItemHeader);
-
 EditableItemHeader.propTypes = {
-  analyticsMetadata,
   afterMenuToggle: PropTypes.func,
+  analyticsMetadata,
   data: PropTypes.object,
-  title: PropTypes.string,
+  Icon: PropTypes.oneOfType([PropTypes.element, PropTypes.node]),
+  maxTitleLength: PropTypes.number,
+  MenuContent: PropTypes.oneOfType([PropTypes.element, PropTypes.node]),
   onTitleChange: PropTypes.func.isRequired,
   priority: PropTypes.number,
-  Icon: PropTypes.oneOfType([
-    PropTypes.element, PropTypes.node,
-  ]),
-  MenuContent: PropTypes.oneOfType([
-    PropTypes.element, PropTypes.node,
-  ]),
-  maxTitleLength: PropTypes.number,
+  title: PropTypes.string,
 };
 
+export default memo(EditableItemHeader);
