@@ -1,25 +1,25 @@
-import React, { forwardRef, useCallback, useEffect, memo } from 'react';
-import { Overlay, Popover } from 'react-bootstrap';
+import React, { forwardRef, memo, useCallback, useEffect } from 'react';
+import Overlay from 'react-bootstrap/Overlay';
+import Popover from 'react-bootstrap/Popover';
+
 import styles from './styles.module.scss';
 
 // eslint-disable-next-line react/display-name
-const FilterSettingsControl = forwardRef((props, ref) => {
-  const { isOpen, hideFilterSettings, target, container, children } = props;
+const FilterSettingsControl = forwardRef(({ isOpen, hideFilterSettings, target, container, children }, ref) => {
+  const handleKeyDown = useCallback((event) => {
+    event.stopPropagation();
+    event.preventDefault();
 
-  const handleKeyDown = useCallback((e) => {
-    e.stopPropagation();
-    e.preventDefault();
-    const { key } = e;
-    if (key === 'Escape') {
+    if (event.key === 'Escape') {
       hideFilterSettings();
     }
-    e.preventDefault();
   }, [hideFilterSettings]);
 
-  const handleOutsideClick = useCallback((e) => {
-    e.stopPropagation();
-    e.preventDefault();
-    if (container.current && (!container.current.contains(e.target))) {
+  const handleOutsideClick = useCallback((event) => {
+    event.stopPropagation();
+    event.preventDefault();
+
+    if (container.current && (!container.current.contains(event.target))) {
       hideFilterSettings();
     }
   }, [container, hideFilterSettings]);
@@ -29,6 +29,7 @@ const FilterSettingsControl = forwardRef((props, ref) => {
       document.addEventListener('mousedown', handleOutsideClick);
       document.addEventListener('keydown', handleKeyDown);
     }
+
     return () => {
       document.removeEventListener('mousedown', handleOutsideClick);
       document.removeEventListener('keydown', handleKeyDown);
@@ -38,17 +39,16 @@ const FilterSettingsControl = forwardRef((props, ref) => {
   return <div>
     <Overlay show={isOpen} target={target.current} container={container.current} placement='bottom' >
       <Popover id="patrol-filter-settings" className={styles.popover}>
-        <Popover.Content ref={ref}>
+        <Popover.Body ref={ref}>
           {children}
-        </Popover.Content>
+        </Popover.Body>
       </Popover>
     </Overlay>
   </div>;
 });
 
-
-export default (memo(FilterSettingsControl));
-
 FilterSettingsControl.defaultProps = {
   defaultSearchSetting: 'start_dates',
 };
+
+export default (memo(FilterSettingsControl));
