@@ -1,29 +1,33 @@
-import React, { useRef, memo } from 'react';
-import { connect } from 'react-redux';
-import { OverlayTrigger, Popover } from 'react-bootstrap';
+import React, { memo, useCallback, useRef } from 'react';
+import OverlayTrigger from 'react-bootstrap/OverlayTrigger';
+import Popover from 'react-bootstrap/Popover';
+import { useSelector } from 'react-redux';
 
-import MapLockControl from '../MapLockControl';
-import MapNamesControl from '../MapNamesControl';
-import UserLocationMapControl from '../UserLocationMapControl';
-import MapDataZoomSimplificationControl from '../MapDataZoomSimplificationControl';
-import MapTrackTimepointsControl from '../MapTrackTimepointsControl';
-import Map3DToggleControl from './Map3DToggleControl';
 import { ReactComponent as GearIcon } from '../common/images/icons/gear.svg';
+
 import { trackEventFactory, MAP_INTERACTION_CATEGORY } from '../utils/analytics';
+
 import ClusterMemberControl from './ClusterMemberControl';
 import InactiveRadioControl from '../InactiveRadioControl';
+import Map3DToggleControl from './Map3DToggleControl';
+import MapDataZoomSimplificationControl from '../MapDataZoomSimplificationControl';
+import MapLockControl from '../MapLockControl';
+import MapNamesControl from '../MapNamesControl';
+import MapTrackTimepointsControl from '../MapTrackTimepointsControl';
+import UserLocationMapControl from '../UserLocationMapControl';
+
 import styles from './styles.module.scss';
 
 const mapInteractionTracker = trackEventFactory(MAP_INTERACTION_CATEGORY);
 
-const MapSettingsControl = (props) => {
-  const { hasUserLocation } = props;
+const MapSettingsControl = () => {
+  const hasUserLocation = useSelector((state) => !!state.view.userLocation);
 
   const formRef = useRef(null);
 
   const popover = (
     <Popover id="settings-popover" title="Map Settings">
-      <Popover.Content>
+      <Popover.Body>
         <ul className={styles.mapSettingsList}>
           <li><MapLockControl /></li>
           <li><MapNamesControl /></li>
@@ -34,24 +38,19 @@ const MapSettingsControl = (props) => {
           <li><Map3DToggleControl /></li>
           {hasUserLocation && <li><UserLocationMapControl /></li>}
         </ul>
-      </Popover.Content>
+      </Popover.Body>
     </Popover>
   );
 
-  const onButtonClick = () => {
+  const onButtonClick = useCallback(() => {
     mapInteractionTracker.track('Clicked Map Settings button');
-  };
+  }, []);
 
-  return <OverlayTrigger trigger="click" placement='left' rootClose={true} overlay={popover}>
-    <button type='button' className={styles.gearButton} ref={formRef}
-      onClick={onButtonClick}>
+  return <OverlayTrigger overlay={popover} placement='left' rootClose trigger="click">
+    <button className={styles.gearButton} onClick={onButtonClick} ref={formRef} type='button'>
       <GearIcon />
     </button>
   </OverlayTrigger>;
 };
 
-const mapStateToProps = ({ view: { userLocation } }) => ({
-  hasUserLocation: !!userLocation,
-});
-
-export default connect(mapStateToProps, null)(memo(MapSettingsControl));
+export default memo(MapSettingsControl);
