@@ -1,31 +1,29 @@
 
 import React, { memo, forwardRef } from 'react';
-import PropTypes from 'prop-types';
 import Popover from 'react-bootstrap/Popover';
-// import Button from 'react-bootstrap/Button';
+import PropTypes from 'prop-types';
+
+import { withFormDataContext } from '../EditableItem/context';
 
 import PriorityPicker from '../PriorityPicker';
-import { withFormDataContext } from '../EditableItem/context';
 
 import styles from './styles.module.scss';
 
-const PatrolHeaderMenuContent = (props, ref) => {
-  const { data, onPrioritySelect, ...rest } = props;
-
+const PatrolHeaderMenuContent = ({ data, onPrioritySelect, ...rest }, ref) => {
   return <Popover {...rest} ref={ref} className={styles.headerPopover}>
-    <Popover.Title>Patrol {data.serial_number}</Popover.Title>
-    <Popover.Content>
+    <Popover.Header>Patrol {data.serial_number}</Popover.Header>
+
+    <Popover.Body>
       <h6>Priority:</h6>
       <PriorityPicker selected={data.priority} onSelect={onPrioritySelect} />
       <hr />
-      {/* <Button type='button' variant='primary'>Export Patrol</Button> */}
-    </Popover.Content>
+    </Popover.Body>
   </Popover>;
 };
-
-export default memo(withFormDataContext(forwardRef(PatrolHeaderMenuContent)));
 
 PatrolHeaderMenuContent.propTypes = {
   data: PropTypes.object.isRequired,
   onPrioritySelect: PropTypes.func.isRequired,
 };
+
+export default memo(withFormDataContext(forwardRef(PatrolHeaderMenuContent)));
