@@ -25,7 +25,7 @@ const UserMenu = (props) => {
 
   return <Fragment>
     <button hidden id="ot-sdk-btn" className="ot-sdk-show-settings" ref={cookieSettingsRef} />
-    <Dropdown alignRight className={styles.menu} {...rest} onToggle={onDropdownToggle}>
+    <Dropdown align="end" className={styles.menu} {...rest} onToggle={onDropdownToggle}>
       <Toggle>
         <UserIcon className={styles.icon} /><span className={styles.username}>{displayUser.username}</span>
       </Toggle>
